physical-ai-book/api/requirements.txt
fastapi
uvicorn
openai-agents
qdrant-client
python-dotenv
pydantic